+++ ESFramework/Core/ESHelpers.m
@@ -224,14 +224,6 @@ NSData *ESRandomDataOfLength(NSUInteger length)
     return [data copy];
 }
 
-UIColor *ESRandomColor(void)
-{
-    return [UIColor colorWithRed:(CGFloat)arc4random() / UINT_MAX
-                           green:(CGFloat)arc4random() / UINT_MAX
-                            blue:(CGFloat)arc4random() / UINT_MAX
-                           alpha:1.0];
-}
-
 CGFloat ESDegreesToRadians(CGFloat degrees)
 {
     return (degrees * M_PI / 180);